requirements.txt
# Core
pydantic>=2.0
pydantic-settings>=2.0
pytz>=2023.0

# Web API
fastapi>=0.100.0
uvicorn>=0.20.0

# News sources
feedparser>=6.0
requests>=2.28
gnews>=0.3.0

# Content extraction
newspaper3k>=0.2.8
lxml_html_clean
beautifulsoup4>=4.12
crawl4ai>=0.2.0

# Geocoding
geopy>=2.3

# AI providers
openai>=1.0
anthropic>=0.20.0

# Database
psycopg2-binary>=2.9

# Browser automation (for crawl4ai)
playwright>=1.30
